+++ src/AutoVer/Services/IO/IPathManager.cs
@@ -2,5 +2,7 @@ namespace AutoVer.Services.IO;
 
 public interface IPathManager
 {
+    char DirectorySeparatorChar { get; }
     string Combine(params string[] paths);
+    string? GetExtension(string? path);
 }